requirements.txt
Django==1.11.6
djangorestframework==3.7.1
requests==2.18.4
